exported Testimony carousel and visitor count to separate jxs file
