+++ src/components/Home.css
@@ -1,148 +1,3 @@
-/* .custom-prev-arrow, .custom-next-arrow {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: rgba(0, 0, 0, 0.5);
-  border-radius: 50%;
-  width: 40px;
-  height: 40px;
-  z-index: 2;
-  cursor: pointer;
-} */
-
-/* .custom-prev-arrow {
-  left: 20px;
-}
-
-.custom-next-arrow {
-  right: 200px;
-} */
-
-.custom-prev-arrow img{
-  position: relative;
-right: 5px;
-}
-.custom-next-arrow img{
-  /* display: none; */
-  position: relative;
-right: -5px;
-}
-
-.custom-prev-arrow img,
-.custom-next-arrow img {
-  width: 35px;
-  height: 35px;
-}
-
-.slick-next{
-  right: -10px !important;
-}
-.slick-prev::before,
-.slick-next::before {
-  display: none;
-  /* left: 2px; */
-  /* color: rgb(8, 10, 52) !important; */
-}
-
-.slick-dots {
-  bottom: -24px !important;
-}
-
-.slick-slide .card {
-  border: none; /* Removes the border from the card element inside the slick slide */
-  box-shadow: 5px 5px;
-  margin-bottom: 20px;
-  padding-left: 15px;
-  text-align: center; /* Centers text content inside the card */
-  display: flex; /* Optionally, align content vertically */
-  flex-direction: column; /* Keeps content stacked */
-  justify-content: center; /* Centers content vertically */
-  align-items: center; /* Centers content horizontally */
-  box-sizing: border-box; /* Prevents padding from increasing card width */
-}
-
-.carouselKarticka .card-img {
-  flex-shrink: 0;
-  width: 150px;
-  height: 150px;
-  margin-top: 20px;
-  margin-bottom: 1rem; /* Add spacing below the image */
-}
-
-@media (max-width: 769px) {
-  .carouselKarticka {
-    flex-direction: column;
-  }
-}
-
-.carouselKarticka {
-  display: flex;
-  align-items: center;
-  text-align: center;
-  gap: 1rem;
-  border: none;
-  min-height: 100%;
-  justify-content: center;
-  flex: 1; /* Ensures the content inside the card grows to fit the height */
-  height: 100%; /* Ensures it stretches to take the same height as the parent card */
-  box-sizing: border-box; /* Prevents padding from increasing the card width */
-}
-
-.card-body {
-  max-width: 400px;
-  min-height: 200px;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center; /* Ensures vertical centering */
-  align-items: center; /* Ensures horizontal centering */
-  text-align: center; /* Centers the text */
-}
-
-@media (max-width: 374px) {
-  .card-body {
-    min-height: 320px; /* Display one card per row on even smaller screens */
-  }
-}
-
-@media (min-width: 767px) and (max-width: 1020px) {
-  .card-body {
-    min-height: 200px; /* Display one card per row on even smaller screens */
-  }
-}
-
-@media (min-width: 1025px) and (max-width: 1439px) {
-  .card-body {
-    min-height: 450px; /* Display one card per row on even smaller screens */
-  }
-}
-
-@media (max-width: 769px) {
-  .carouselKarticka .card-img {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    text-align: center;
-  }
-}
-
-.card-body h5,
-.card-body h6 {
-  margin: 0.5rem 0; /* Adjust spacing for text */
-}
-
-.Karousel {
-  max-width: 300px;
-}
-
-.slider-container {
-  margin-left: 1.2rem;
-  margin-right: 1.2rem;
-  margin-bottom: 1.5rem;
-}
-.slider-container .slick-slide {
-  padding: 0 15px; /* Adds more space between each slide */
-}
 
 .background {
   background-image: url('/public/images/BGImage2.jpg');
@@ -257,9 +112,10 @@ right: -5px;
 }
 
 /* Arrow styling */
+
 /* .carousel-control-prev::before,
 .carousel-control-next::before {
-  content: ""; 
+  content: "";
   width: 0;
   height: 0;
   border-style: solid;
